requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
chardet==3.0.4
discord.py==1.2.5
google==2.0.2
idna==2.8
multidict==4.6.1
psycopg2==2.8.4
python-dotenv==0.10.3
redis==3.3.11
soupsieve==1.9.5
websockets==6.0
yarl==1.3.0

